+++ js/utils.js
@@ -1,0 +1,8 @@
+
+const getRandomNumber = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
+
+//const checkCommentSize = (checkedString, maxLength) => checkedString.length <= maxLength;
+
+const getRandomArrayElement = (elements) => elements[getRandomNumber(0, elements.length - 1)];
+
+export {getRandomArrayElement, getRandomNumber};
